refactor #24: Layout router 활용하도록 App 구조변경
